Move MobileWork files to relevant WO or WOs folder
