+++ src/components/WorkOrders/MobileWorkingWorkOrdersView.js
@@ -4,7 +4,7 @@ import { beginningOfDay } from '@/utils/time'
 import { longMonthWeekday } from '@/utils/date'
 import Spinner from '../Spinner'
 import ErrorMessage from '../Errors/ErrorMessage'
-import MobileWorkingWorkOrderListItem from './MobileWorkingWorkOrderListItem'
+import MobileWorkingWorkOrderListItem from '../WorkOrder/MobileWorkingWorkOrderListItem'
 import WarningInfoBox from '../Template/WarningInfoBox'
 import Meta from '../Meta'
 import { WorkOrder } from '../../models/workOrder'
